Restructure Test files for easier scripting
This changes up the directory structure to be a little easier to script.
It also allows for the tests to be named (via the name of the directory)
instead of strictly numbered

In the new structure we have a directory under `tests` for each test.
Currently they are all numbered but they could easily be named instead.
Inside each of these directories is an `input.json` and `output.json`
file.